+++ docs/court_wizard.d.ts
@@ -6,14 +6,14 @@ export type InitInput = RequestInfo | URL | Response | BufferSource | WebAssembl
 export interface InitOutput {
   readonly memory: WebAssembly.Memory;
   readonly main: (a: number, b: number) => number;
-  readonly __wasm_bindgen_func_elem_125178: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_125157: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_129864: (a: number, b: number, c: number) => void;
-  readonly __wasm_bindgen_func_elem_129742: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_128205: (a: number, b: number, c: number) => void;
-  readonly __wasm_bindgen_func_elem_128195: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_129863: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_129865: (a: number, b: number, c: number, d: number) => void;
+  readonly __wasm_bindgen_func_elem_125195: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_125174: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_128222: (a: number, b: number, c: number) => void;
+  readonly __wasm_bindgen_func_elem_128212: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_129881: (a: number, b: number, c: number) => void;
+  readonly __wasm_bindgen_func_elem_129759: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_129880: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_129882: (a: number, b: number, c: number, d: number) => void;
   readonly __wbindgen_export: (a: number, b: number) => number;
   readonly __wbindgen_export2: (a: number, b: number, c: number, d: number) => number;
   readonly __wbindgen_export3: (a: number) => void;