+++ docs/court_wizard.js
@@ -1,4 +1,4 @@
-import { copy_to_clipboard } from './snippets/court_wizard-5a16a977edd52ccc/inline0.js';
+import { copy_to_clipboard } from './snippets/court_wizard-deca32bb11134a69/inline0.js';
 
 const lAudioContext = (typeof AudioContext !== 'undefined' ? AudioContext : (typeof webkitAudioContext !== 'undefined' ? webkitAudioContext : undefined));
 let wasm;
@@ -258,24 +258,24 @@ if (!('encodeInto' in cachedTextEncoder)) {
 
 let WASM_VECTOR_LEN = 0;
 
-function __wasm_bindgen_func_elem_125178(arg0, arg1) {
-    wasm.__wasm_bindgen_func_elem_125178(arg0, arg1);
+function __wasm_bindgen_func_elem_125195(arg0, arg1) {
+    wasm.__wasm_bindgen_func_elem_125195(arg0, arg1);
 }
 
-function __wasm_bindgen_func_elem_129864(arg0, arg1, arg2) {
-    wasm.__wasm_bindgen_func_elem_129864(arg0, arg1, addHeapObject(arg2));
+function __wasm_bindgen_func_elem_128222(arg0, arg1, arg2) {
+    wasm.__wasm_bindgen_func_elem_128222(arg0, arg1, addHeapObject(arg2));
 }
 
-function __wasm_bindgen_func_elem_128205(arg0, arg1, arg2) {
-    wasm.__wasm_bindgen_func_elem_128205(arg0, arg1, addHeapObject(arg2));
+function __wasm_bindgen_func_elem_129881(arg0, arg1, arg2) {
+    wasm.__wasm_bindgen_func_elem_129881(arg0, arg1, addHeapObject(arg2));
 }
 
-function __wasm_bindgen_func_elem_129863(arg0, arg1) {
-    wasm.__wasm_bindgen_func_elem_129863(arg0, arg1);
+function __wasm_bindgen_func_elem_129880(arg0, arg1) {
+    wasm.__wasm_bindgen_func_elem_129880(arg0, arg1);
 }
 
-function __wasm_bindgen_func_elem_129865(arg0, arg1, arg2, arg3) {
-    wasm.__wasm_bindgen_func_elem_129865(arg0, arg1, addHeapObject(arg2), addHeapObject(arg3));
+function __wasm_bindgen_func_elem_129882(arg0, arg1, arg2, arg3) {
+    wasm.__wasm_bindgen_func_elem_129882(arg0, arg1, addHeapObject(arg2), addHeapObject(arg3));
 }
 
 const __wbindgen_enum_GpuAddressMode = ["clamp-to-edge", "repeat", "mirror-repeat"];
@@ -604,7 +604,7 @@ function __wbg_get_imports() {
     imports.wbg.__wbg_copyToChannel_30d90303302ec449 = function() { return handleError(function (arg0, arg1, arg2, arg3) {
         getObject(arg0).copyToChannel(getArrayF32FromWasm0(arg1, arg2), arg3);
     }, arguments) };
-    imports.wbg.__wbg_copy_to_clipboard_4fabb5931c022aa2 = function(arg0, arg1) {
+    imports.wbg.__wbg_copy_to_clipboard_7de6f8a6322da958 = function(arg0, arg1) {
         copy_to_clipboard(getStringFromWasm0(arg0, arg1));
     };
     imports.wbg.__wbg_createAnswer_57fa5e0880a7b92a = function(arg0) {
@@ -2349,19 +2349,9 @@ function __wbg_get_imports() {
         const ret = getObject(arg0).y;
         return ret;
     };
-    imports.wbg.__wbindgen_cast_01237dc173bbd89c = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 85866, function: Function { arguments: [NamedExternref("Array<any>")], shim_idx: 85868, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_129742, __wasm_bindgen_func_elem_129864);
-        return addHeapObject(ret);
-    };
-    imports.wbg.__wbindgen_cast_0f67690c28d1de92 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 85866, function: Function { arguments: [NamedExternref("Event")], shim_idx: 85868, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_129742, __wasm_bindgen_func_elem_129864);
-        return addHeapObject(ret);
-    };
-    imports.wbg.__wbindgen_cast_1c9f8655fca7d0ea = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 85866, function: Function { arguments: [NamedExternref("KeyboardEvent")], shim_idx: 85868, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_129742, __wasm_bindgen_func_elem_129864);
+    imports.wbg.__wbindgen_cast_17485b13ebf34b16 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 85879, function: Function { arguments: [NamedExternref("Event")], shim_idx: 85881, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_129759, __wasm_bindgen_func_elem_129881);
         return addHeapObject(ret);
     };
     imports.wbg.__wbindgen_cast_2241b6af4c4b2941 = function(arg0, arg1) {
@@ -2369,39 +2359,49 @@ function __wbg_get_imports() {
         const ret = getStringFromWasm0(arg0, arg1);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_456cc9ab5c277cfa = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 85808, function: Function { arguments: [Externref], shim_idx: 85809, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_128195, __wasm_bindgen_func_elem_128205);
+    imports.wbg.__wbindgen_cast_32cf302b87727dce = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 85879, function: Function { arguments: [NamedExternref("WheelEvent")], shim_idx: 85881, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_129759, __wasm_bindgen_func_elem_129881);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_4b71b45afb5c5871 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 85866, function: Function { arguments: [NamedExternref("PageTransitionEvent")], shim_idx: 85868, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_129742, __wasm_bindgen_func_elem_129864);
+    imports.wbg.__wbindgen_cast_4fc49e9adfc0103e = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 85879, function: Function { arguments: [NamedExternref("Array<any>"), NamedExternref("ResizeObserver")], shim_idx: 85882, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_129759, __wasm_bindgen_func_elem_129882);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_64b4dcb55ec14b9d = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 85866, function: Function { arguments: [NamedExternref("PointerEvent")], shim_idx: 85868, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_129742, __wasm_bindgen_func_elem_129864);
+    imports.wbg.__wbindgen_cast_509f4d0070f495ba = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 85250, function: Function { arguments: [], shim_idx: 85251, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125174, __wasm_bindgen_func_elem_125195);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_98002ea304c1fdcc = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 85237, function: Function { arguments: [], shim_idx: 85238, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125157, __wasm_bindgen_func_elem_125178);
+    imports.wbg.__wbindgen_cast_9d0cae517d3b6a0f = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 85821, function: Function { arguments: [Externref], shim_idx: 85822, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_128212, __wasm_bindgen_func_elem_128222);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_9e6874aeca0a882d = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 85866, function: Function { arguments: [], shim_idx: 85867, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_129742, __wasm_bindgen_func_elem_129863);
+    imports.wbg.__wbindgen_cast_a31ba99d63a597ce = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 85879, function: Function { arguments: [NamedExternref("PointerEvent")], shim_idx: 85881, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_129759, __wasm_bindgen_func_elem_129881);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_a198a58ead7e3b72 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 85866, function: Function { arguments: [NamedExternref("WheelEvent")], shim_idx: 85868, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_129742, __wasm_bindgen_func_elem_129864);
+    imports.wbg.__wbindgen_cast_b3152158490a2918 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 85879, function: Function { arguments: [NamedExternref("FocusEvent")], shim_idx: 85881, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_129759, __wasm_bindgen_func_elem_129881);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_bad8e3ee0f8f0708 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 85866, function: Function { arguments: [NamedExternref("Array<any>"), NamedExternref("ResizeObserver")], shim_idx: 85869, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_129742, __wasm_bindgen_func_elem_129865);
+    imports.wbg.__wbindgen_cast_b45a11ea60eb3afc = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 85879, function: Function { arguments: [NamedExternref("Array<any>")], shim_idx: 85881, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_129759, __wasm_bindgen_func_elem_129881);
+        return addHeapObject(ret);
+    };
+    imports.wbg.__wbindgen_cast_be73550d54d04122 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 85879, function: Function { arguments: [NamedExternref("PageTransitionEvent")], shim_idx: 85881, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_129759, __wasm_bindgen_func_elem_129881);
+        return addHeapObject(ret);
+    };
+    imports.wbg.__wbindgen_cast_c4c9e4105d3fc634 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 85879, function: Function { arguments: [NamedExternref("KeyboardEvent")], shim_idx: 85881, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_129759, __wasm_bindgen_func_elem_129881);
         return addHeapObject(ret);
     };
     imports.wbg.__wbindgen_cast_cb9088102bce6b30 = function(arg0, arg1) {
@@ -2409,14 +2409,14 @@ function __wbg_get_imports() {
         const ret = getArrayU8FromWasm0(arg0, arg1);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_d564869292089e59 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 85866, function: Function { arguments: [NamedExternref("FocusEvent")], shim_idx: 85868, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_129742, __wasm_bindgen_func_elem_129864);
-        return addHeapObject(ret);
-    };
     imports.wbg.__wbindgen_cast_d6cd19b81560fd6e = function(arg0) {
         // Cast intrinsic for `F64 -> Externref`.
         const ret = arg0;
+        return addHeapObject(ret);
+    };
+    imports.wbg.__wbindgen_cast_e997cd4d39c8b9d6 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 85879, function: Function { arguments: [], shim_idx: 85880, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_129759, __wasm_bindgen_func_elem_129880);
         return addHeapObject(ret);
     };
     imports.wbg.__wbindgen_object_clone_ref = function(arg0) {