+++ docs/court_wizard_bg.wasm.d.ts
@@ -2,14 +2,14 @@
 /* eslint-disable */
 export const memory: WebAssembly.Memory;
 export const main: (a: number, b: number) => number;
-export const __wasm_bindgen_func_elem_125178: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_125157: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_129864: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_129742: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_128205: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_128195: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_129863: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_129865: (a: number, b: number, c: number, d: number) => void;
+export const __wasm_bindgen_func_elem_125195: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_125174: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_128222: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_128212: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_129881: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_129759: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_129880: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_129882: (a: number, b: number, c: number, d: number) => void;
 export const __wbindgen_export: (a: number, b: number) => number;
 export const __wbindgen_export2: (a: number, b: number, c: number, d: number) => number;
 export const __wbindgen_export3: (a: number) => void;